sm6150-common: properties: Disable vsync for CPU rendered Apps
* Choreographer VSync has some performance impact for CPU rendered apps
as the app has to wait for VSync to start a new frame draw. When the
frame draw time exceeds one VSync, this has a direct impact on the FPS.
debug.cpurend.vsync property controls the VSync for CPU rendered apps.

Disable VSync for CPU rendered app, by setting this property to false.

Change-Id: If8c6f9e182eac6d015dbb69fe2a74fc9b3a85c1f
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>
Signed-off-by: adawong <orcalesyuuno@gmail.com>